pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "sql-debug-env"
version = "1.0.0"
description = "An OpenEnv-compliant RL environment for debugging broken SQL queries"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
keywords = ["openenv", "reinforcement-learning", "sql", "agent", "evaluation"]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "aiohttp>=3.9.0",
    "openai>=1.12.0",
    "openenv-core>=0.2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27"]

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
